Theming: Add support for optional light action and system bars on day themes:
Settings:
- Optional light status and action bar on day themes
- Optional light navigation bar on day themes
Terminal:
- Optional light status and action bar on day themes

For now only use the optional light navigation bar on settings,
the terminal background color is customizeable,
so even on a day theme the background color can be set to maybe black,
or yellow, ... so a light grey navigation bar could be looking ugly.
I didn't decided how to handle these scenarios, so maybe
support for the optional light navigation bar will be added
later to the terminal itelsf.