Fix right boarder and hide scrollbar for now:
It seems the terminal background doesn't fit the full
width of the TerminalView, which results in a small
boarder on the right side.
Set the TerminalView background color to the terminal background color,
so the boarder is 'gone'.
Also hide the TerminalView scollbar for now.
The scollbar color should be related to the terminal background color,
instead of related to the currently applied theme.

TO-DO:
Use a custom scrollbar, to make colorizing more easy.